requirements.txt
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
beautifulsoup4==4.6.0
boto3==1.12.34
botocore==1.15.34
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==3.0.3
django-debug-toolbar==2.1
django-extensions==2.2.7
django-modelcluster==5.0.1
django-recaptcha==2.0.6
django-storages==1.9.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.8
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
jmespath==0.9.5
l18n==2018.5
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.8
wagtail-django-recaptcha==1.0
wcwidth==0.1.8
webencodings==0.5.1
Willow==1.3
